Clean up dead code, redundant casts, and compiler warnings
- Remove dead if/else in sgb.rs where both branches were identical
- Remove unused is_vectorize variable in main.rs
- Remove unused SDL_Camera import in main.rs
- Fix redundant double-casts ((val & 0x07) as u16) as u32 in apu.rs
- Fix redundant as u16 as f32 casts in main.rs accelerometer constants
- Use stack array instead of heap vec for SGB VRAM transfer buffer
- Use VecDeque instead of Vec with remove(0) in trace_timer command
- Use .fill(0) instead of manual iterator loop in PPU VRAM init
- Fix unused_mut and unused variable warnings across codebase

+++ src/apu.rs
@@ -897,7 +897,7 @@ impl Apu {
                 // did_tick frequency change edge case (must check BEFORE freq update)
                 if val & 0x80 == 0 && self.ch1.enabled {
                     let old_hi = (old_freq >> 8) & 7;
-                    let new_hi = ((val & 0x07) as u16) as u32;
+                    let new_hi = (val & 0x07) as u32;
                     if old_hi == 7 && new_hi != 7 && self.ch1.did_tick {
                         if (self.ch1.freq_timer.wrapping_sub(2)) / 4 == (old_freq ^ 0x7FF) as u32 {
                             self.ch1.duty_pos = self.ch1.duty_pos.wrapping_sub(1) & 7;
@@ -950,7 +950,7 @@ impl Apu {
                 // did_tick frequency change edge case (must check BEFORE freq update)
                 if val & 0x80 == 0 && self.ch2.enabled {
                     let old_hi = (old_freq >> 8) & 7;
-                    let new_hi = ((val & 0x07) as u16) as u32;
+                    let new_hi = (val & 0x07) as u32;
                     if old_hi == 7 && new_hi != 7 && self.ch2.did_tick {
                         if (self.ch2.freq_timer.wrapping_sub(2)) / 4 == (old_freq ^ 0x7FF) as u32 {
                             self.ch2.duty_pos = self.ch2.duty_pos.wrapping_sub(1) & 7;
